@@ -15,21 +15,21 @@
     "prepublishOnly": "yarn build"
   },
   "dependencies": {
-    "@mainframe/rpc-browser": "^0.1.0",
+    "@mainframe/rpc-browser": "^0.2.0",
     "@mainframe/utils-hex": "^0.1.0",
     "erebos-api-bzz-browser": "*",
     "erebos-api-eth": "*",
     "erebos-api-net": "*",
     "erebos-api-pss": "*",
-    "erebos-api-pss": "*",
+    "erebos-api-shh": "*",
     "erebos-api-web3": "*",
     "erebos-client-base": "*",
-    "rxjs": "^5.5.9"
+    "rxjs": "^6.0.0"
   },
   "devDependencies": {
     "babel-loader": "^7.1.4",
-    "flow-bin": "^0.69.0",
-    "webpack": "^4.5.0",
-    "webpack-cli": "^2.0.14"
+    "flow-bin": "^0.70.0",
+    "webpack": "^4.6.0",
+    "webpack-cli": "^2.0.15"
   }
 }
