@@ -1,5 +1,5 @@
-// flow-typed signature: 12eedcdb3fb0550d9b4f6227313ac585
-// flow-typed version: <<STUB>>/@angular-devkit/build-optimizer_v0.0.24/flow_v0.69.0
+// flow-typed signature: 983bef359570886009e638b0f59bc44a
+// flow-typed version: <<STUB>>/@angular-devkit/build-optimizer_v0.4.6/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
