@@ -1,10 +1,10 @@
-// flow-typed signature: e1171bad33a469e6da7305891160e157
-// flow-typed version: <<STUB>>/sinon_v4.3.0/flow_v0.70.0
+// flow-typed signature: dc439562a6481ef3728b7615b80dc7ad
+// flow-typed version: <<STUB>>/@angular-devkit/schematics_v^0.5.4/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
  *
- *   'sinon'
+ *   '@angular-devkit/schematics'
  *
  * Fill this stub out by replacing all the `any` types.
  *
@@ -13,6 +13,6 @@
  * https://github.com/flowtype/flow-typed
  */
 
-declare module 'sinon' {
+declare module '@angular-devkit/schematics' {
   declare module.exports: any;
 }
