@@ -1,5 +1,5 @@
-// flow-typed signature: 84e0307aeac22d6612678140d79ba222
-// flow-typed version: <<STUB>>/@mainframe/rpc-base_v^0.1.0/flow_v0.69.0
+// flow-typed signature: 9c09f86312d5c7157f4da65a6134d3e4
+// flow-typed version: <<STUB>>/@mainframe/rpc-base_v^0.1.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
