@@ -1,5 +1,5 @@
-// flow-typed signature: 68e3cde02204dfb8c9edaa128c673daf
-// flow-typed version: <<STUB>>/@mainframe/rpc-http-browser_v^0.1.0/flow_v0.69.0
+// flow-typed signature: dfd134b7a6971379e4fc524a1a1257ac
+// flow-typed version: <<STUB>>/@mainframe/rpc-http-browser_v^0.1.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
