@@ -1,5 +1,5 @@
-// flow-typed signature: c163c0e3850f008a5f44d67c0816c591
-// flow-typed version: <<STUB>>/@mainframe/rpc-http-node_v^0.1.0/flow_v0.69.0
+// flow-typed signature: cebebcd8029bcccccbbc56fbe6a8808a
+// flow-typed version: <<STUB>>/@mainframe/rpc-http-node_v^0.1.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
