@@ -1,5 +1,5 @@
-// flow-typed signature: c331aa13ee2430e264e5db4e20e860e1
-// flow-typed version: <<STUB>>/@mainframe/rpc-ipc_v^0.1.0/flow_v0.69.0
+// flow-typed signature: 9d640664afc26f181c128752d9a93213
+// flow-typed version: <<STUB>>/@mainframe/rpc-ipc_v^0.2.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
