@@ -1,5 +1,5 @@
-// flow-typed signature: 036be4b0624a2d60d1ac797612d74564
-// flow-typed version: <<STUB>>/@mainframe/rpc-web3_v^0.1.0/flow_v0.69.0
+// flow-typed signature: 69582ec171a126e8b49aa55f7fc789f8
+// flow-typed version: <<STUB>>/@mainframe/rpc-web3_v^0.1.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
