@@ -1,5 +1,5 @@
-// flow-typed signature: ac70b468d741561619af3571f30b4b9e
-// flow-typed version: <<STUB>>/@mainframe/rpc-ws-browser_v^0.1.0/flow_v0.69.0
+// flow-typed signature: f58dfd2878cfb300a035593f39b7b76e
+// flow-typed version: <<STUB>>/@mainframe/rpc-ws-browser_v^0.2.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
