@@ -1,5 +1,5 @@
-// flow-typed signature: 9e2c9f5e30202020a1bb7016709da0fc
-// flow-typed version: <<STUB>>/@mainframe/rpc-ws-node_v^0.1.0/flow_v0.69.0
+// flow-typed signature: fa40944b719806da4dd109cde773d9c1
+// flow-typed version: <<STUB>>/@mainframe/rpc-ws-node_v^0.2.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
