@@ -1,10 +1,10 @@
-// flow-typed signature: e1171bad33a469e6da7305891160e157
-// flow-typed version: <<STUB>>/sinon_v4.3.0/flow_v0.70.0
+// flow-typed signature: 008371d63cc625e7eba814d12ae0408b
+// flow-typed version: <<STUB>>/@types/chai_v4.1.2/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
  *
- *   'sinon'
+ *   '@types/chai'
  *
  * Fill this stub out by replacing all the `any` types.
  *
@@ -13,6 +13,6 @@
  * https://github.com/flowtype/flow-typed
  */
 
-declare module 'sinon' {
+declare module '@types/chai' {
   declare module.exports: any;
 }
