@@ -1,10 +1,10 @@
-// flow-typed signature: fe1bed9a494fa6bd8d93f91d966dfd9a
-// flow-typed version: <<STUB>>/npm-scripts-info_v^0.3.4/flow_v0.69.0
+// flow-typed signature: bf035eaf271672de005e18e79001c61e
+// flow-typed version: <<STUB>>/@types/lodash_v4.14.102/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
  *
- *   'npm-scripts-info'
+ *   '@types/lodash'
  *
  * Fill this stub out by replacing all the `any` types.
  *
@@ -13,6 +13,6 @@
  * https://github.com/flowtype/flow-typed
  */
 
-declare module 'npm-scripts-info' {
+declare module '@types/lodash' {
   declare module.exports: any;
 }
