@@ -1,10 +1,10 @@
-// flow-typed signature: e1171bad33a469e6da7305891160e157
-// flow-typed version: <<STUB>>/sinon_v4.3.0/flow_v0.70.0
+// flow-typed signature: 3c2c6b797cb64d6175fae99f199e04a8
+// flow-typed version: <<STUB>>/@types/mocha_v2.2.48/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
  *
- *   'sinon'
+ *   '@types/mocha'
  *
  * Fill this stub out by replacing all the `any` types.
  *
@@ -13,6 +13,6 @@
  * https://github.com/flowtype/flow-typed
  */
 
-declare module 'sinon' {
+declare module '@types/mocha' {
   declare module.exports: any;
 }
