@@ -1,10 +1,10 @@
-// flow-typed signature: 9d7738b4ba186d2adec00ef24b4af4cf
-// flow-typed version: <<STUB>>/typings_v^2.0.0/flow_v0.69.0
+// flow-typed signature: bf83857695a9ec78abaa96f2efa29407
+// flow-typed version: <<STUB>>/@types/node_v9.4.5/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
  *
- *   'typings'
+ *   '@types/node'
  *
  * Fill this stub out by replacing all the `any` types.
  *
@@ -13,6 +13,6 @@
  * https://github.com/flowtype/flow-typed
  */
 
-declare module 'typings' {
+declare module '@types/node' {
   declare module.exports: any;
 }
