@@ -1,10 +1,10 @@
-// flow-typed signature: e1171bad33a469e6da7305891160e157
-// flow-typed version: <<STUB>>/sinon_v4.3.0/flow_v0.70.0
+// flow-typed signature: eb0b77fb924e86d38008a03decc4f79f
+// flow-typed version: <<STUB>>/@types/sinon_v4.1.3/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
  *
- *   'sinon'
+ *   '@types/sinon'
  *
  * Fill this stub out by replacing all the `any` types.
  *
@@ -13,6 +13,6 @@
  * https://github.com/flowtype/flow-typed
  */
 
-declare module 'sinon' {
+declare module '@types/sinon' {
   declare module.exports: any;
 }
