@@ -1,5 +1,5 @@
-// flow-typed signature: 08534359639299d7e7cdc9fed6b0a3b6
-// flow-typed version: <<STUB>>/babel-polyfill_v^6.23.0/flow_v0.69.0
+// flow-typed signature: d418dfa14869dbe369ab3c3a8184b436
+// flow-typed version: <<STUB>>/babel-polyfill_v6.26.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
