@@ -1,5 +1,5 @@
-// flow-typed signature: 43023cbb429ef51f523f906d49eb0c06
-// flow-typed version: <<STUB>>/benchmark_v^2.1.0/flow_v0.69.0
+// flow-typed signature: 334c56e2b5c4f23644f576054f756107
+// flow-typed version: <<STUB>>/benchmark_v2.1.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
