@@ -1,5 +1,5 @@
-// flow-typed signature: c37b129c9689ad02176656147e0665a0
-// flow-typed version: <<STUB>>/benchpress_v2.0.0-beta.1/flow_v0.69.0
+// flow-typed signature: 2f596543203e25c58894bcde80cc0507
+// flow-typed version: <<STUB>>/benchpress_v2.0.0-beta.1/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
