@@ -1,5 +1,5 @@
-// flow-typed signature: 01d5fcb334589809326a1ab2001c7a1c
-// flow-typed version: <<STUB>>/chai_v^3.5.0/flow_v0.69.0
+// flow-typed signature: 92bac57e3d08303dd419ef257c62fe1a
+// flow-typed version: <<STUB>>/chai_v4.1.2/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
