@@ -1,5 +1,5 @@
-// flow-typed signature: 536b63f2e851e7f1a32dfd756922ef88
-// flow-typed version: <<STUB>>/color_v^0.11.1/flow_v0.69.0
+// flow-typed signature: 0fcd4edb143190e871019939a0105ea9
+// flow-typed version: <<STUB>>/color_v3.0.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
