@@ -1,5 +1,5 @@
-// flow-typed signature: 21b5ca13eeb664729dd07e8fa4d92c80
-// flow-typed version: <<STUB>>/colors_v1.1.2/flow_v0.69.0
+// flow-typed signature: 428069145499966a73286b963adb4678
+// flow-typed version: <<STUB>>/colors_v1.1.2/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
