@@ -1,5 +1,5 @@
-// flow-typed signature: 31689f5d0c0731c3f2512a0e8657a440
-// flow-typed version: <<STUB>>/commitizen_v^2.8.6/flow_v0.69.0
+// flow-typed signature: 27115350f21cb3c4c82742f4554a58cb
+// flow-typed version: <<STUB>>/commitizen_v2.9.6/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
