@@ -1,5 +1,5 @@
-// flow-typed signature: 69f00a3e2740c6bb994fcd647a1691d6
-// flow-typed version: <<STUB>>/coveralls_v^2.11.13/flow_v0.69.0
+// flow-typed signature: e8fa8adfdbbfea59dbc8bca49ff6ef20
+// flow-typed version: <<STUB>>/coveralls_v3.0.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
