@@ -1,5 +1,5 @@
-// flow-typed signature: ba4e995004990e4ef2e154e9a3f70f1f
-// flow-typed version: <<STUB>>/cz-conventional-changelog_v^1.2.0/flow_v0.69.0
+// flow-typed signature: e4915af3c53f79e4cb5015b360a18c20
+// flow-typed version: <<STUB>>/cz-conventional-changelog_v1.2.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
