@@ -1,5 +1,5 @@
-// flow-typed signature: f363b2f43e30d3c28fcb42f55bd2ae7a
-// flow-typed version: <<STUB>>/danger_v^1.1.0/flow_v0.69.0
+// flow-typed signature: ee1314975f3a59839f5db8fb025b136c
+// flow-typed version: <<STUB>>/danger_v1.1.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
