@@ -1,10 +1,10 @@
-// flow-typed signature: 52f3799192398ff5a04ed09891843023
-// flow-typed version: <<STUB>>/madge_v^1.4.3/flow_v0.69.0
+// flow-typed signature: a3aa84d075b248d2b16c64b5e08d18d5
+// flow-typed version: <<STUB>>/dependency-cruiser_v2.13.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
  *
- *   'madge'
+ *   'dependency-cruiser'
  *
  * Fill this stub out by replacing all the `any` types.
  *
@@ -13,6 +13,6 @@
  * https://github.com/flowtype/flow-typed
  */
 
-declare module 'madge' {
+declare module 'dependency-cruiser' {
   declare module.exports: any;
 }
