@@ -1,5 +1,5 @@
-// flow-typed signature: 6686ddb5c4939e4ad1bbc32260b8bdae
-// flow-typed version: <<STUB>>/doctoc_v^1.0.0/flow_v0.69.0
+// flow-typed signature: 78a7f205613367b8d1ca583ff4d2006f
+// flow-typed version: <<STUB>>/doctoc_v1.3.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
