@@ -1,5 +1,5 @@
-// flow-typed signature: f2be8eba511c7ef18180374937534da0
-// flow-typed version: <<STUB>>/escape-string-regexp_v^1.0.5 /flow_v0.69.0
+// flow-typed signature: adfe149f111584aa78c3cdf8c305fc30
+// flow-typed version: <<STUB>>/escape-string-regexp_v1.0.5/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
