@@ -1,5 +1,5 @@
-// flow-typed signature: 589305ba077758a844ffb499a7bd6ad3
-// flow-typed version: <<STUB>>/esdoc_v^0.4.7/flow_v0.69.0
+// flow-typed signature: d1fdfcfcfc1d7b32d8dbf0ea3e1aa130
+// flow-typed version: <<STUB>>/esdoc_v0.4.7/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
