@@ -1,5 +1,5 @@
-// flow-typed signature: 3f1fe3aceb2f18833fd4fcff2324d241
-// flow-typed version: <<STUB>>/eslint_v^3.8.0/flow_v0.69.0
+// flow-typed signature: 40bd259248c25fec222d6b0e36a55427
+// flow-typed version: <<STUB>>/eslint_v4.17.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
