@@ -1,5 +1,5 @@
-// flow-typed signature: 17ca18dc7b3a00b8e0771156a84efcb5
-// flow-typed version: <<STUB>>/flow-bin_v^0.69.0/flow_v0.69.0
+// flow-typed signature: aa1642cc1e346823f74c01be21b0fb98
+// flow-typed version: <<STUB>>/flow-bin_v^0.70.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
