@@ -1,5 +1,5 @@
-// flow-typed signature: 52641ee59466ba1ab93f737390d86a95
-// flow-typed version: <<STUB>>/fs-extra_v^2.1.2/flow_v0.69.0
+// flow-typed signature: a837f89446a25f2234373ecbce41a28e
+// flow-typed version: <<STUB>>/fs-extra_v5.0.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
