@@ -1,5 +1,5 @@
-// flow-typed signature: 14f7a25aefa21e36ee9fc828cd9bba90
-// flow-typed version: <<STUB>>/get-folder-size_v^1.0.0/flow_v0.69.0
+// flow-typed signature: d67e87f6910fc713346b6c9e733b4ffc
+// flow-typed version: <<STUB>>/get-folder-size_v1.0.1/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
