@@ -1,5 +1,5 @@
-// flow-typed signature: 4f2432af00c6a58837b679bd210f913a
-// flow-typed version: <<STUB>>/glob_v^7.0.3/flow_v0.69.0
+// flow-typed signature: 2a3f8bb08f1be66fe2c3f442cd7f5c49
+// flow-typed version: <<STUB>>/glob_v7.1.2/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
