@@ -1,5 +1,5 @@
-// flow-typed signature: 7dc65161ac68def1f63e9f9fa3de6e8f
-// flow-typed version: <<STUB>>/gm_v^1.22.0/flow_v0.69.0
+// flow-typed signature: cc75f3c946ba30762f9ed249f3a7da39
+// flow-typed version: <<STUB>>/gm_v1.23.1/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
