@@ -1,5 +1,5 @@
-// flow-typed signature: 52dedb9b47db224ae6ab9db3c779502c
-// flow-typed version: <<STUB>>/google-closure-compiler-js_v^20170218.0.0/flow_v0.69.0
+// flow-typed signature: 00097695871225e5c70cd50d7b7b5456
+// flow-typed version: <<STUB>>/google-closure-compiler-js_v20170218.0.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
