@@ -1,5 +1,5 @@
-// flow-typed signature: 7773b5e1c2e1927efe6f1102113c6e48
-// flow-typed version: <<STUB>>/gzip-size_v^3.0.0/flow_v0.69.0
+// flow-typed signature: 04efa48caf7c58669a739dd62c95e5e1
+// flow-typed version: <<STUB>>/gzip-size_v4.1.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
