@@ -1,5 +1,5 @@
-// flow-typed signature: e85d15c69c9936dbc690194473b1ad87
-// flow-typed version: <<STUB>>/husky_v^0.13.3/flow_v0.69.0
+// flow-typed signature: aec9eccd2b5a66187560b4e81e461074
+// flow-typed version: <<STUB>>/husky_v0.14.3/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
