@@ -1,5 +1,5 @@
-// flow-typed signature: b54a13a7825c14f857b031112a08903a
-// flow-typed version: <<STUB>>/klaw-sync_v^3.0.0/flow_v0.69.0
+// flow-typed signature: 1fc1b61c673a8f4bed703b56a671b2d7
+// flow-typed version: <<STUB>>/klaw-sync_v3.0.2/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
