@@ -1,5 +1,5 @@
-// flow-typed signature: 310741e3bde58835889c74bf542f1162
-// flow-typed version: <<STUB>>/lint-staged_v3.2.5/flow_v0.69.0
+// flow-typed signature: fc9beeb65f6b8e14ca3b6f14062fa01a
+// flow-typed version: <<STUB>>/lint-staged_v3.2.5/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
