@@ -1,5 +1,5 @@
-// flow-typed signature: f39fe4a7677d727b2f40b1ebd3b48fa9
-// flow-typed version: <<STUB>>/lodash_v^4.15.0/flow_v0.69.0
+// flow-typed signature: 7b4a4dc5b619e0fbcd7055d386d7ac4c
+// flow-typed version: <<STUB>>/lodash_v4.17.5/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,14 +22,6 @@ declare module 'lodash' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-declare module 'lodash/_addMapEntry' {
-  declare module.exports: any;
-}
-
-declare module 'lodash/_addSetEntry' {
-  declare module.exports: any;
-}
-
 declare module 'lodash/_apply' {
   declare module.exports: any;
 }
@@ -570,15 +562,7 @@ declare module 'lodash/_cloneDataView' {
   declare module.exports: any;
 }
 
-declare module 'lodash/_cloneMap' {
-  declare module.exports: any;
-}
-
 declare module 'lodash/_cloneRegExp' {
-  declare module.exports: any;
-}
-
-declare module 'lodash/_cloneSet' {
   declare module.exports: any;
 }
 
@@ -1091,6 +1075,10 @@ declare module 'lodash/_replaceHolders' {
 }
 
 declare module 'lodash/_root' {
+  declare module.exports: any;
+}
+
+declare module 'lodash/_safeGet' {
   declare module.exports: any;
 }
 
@@ -4215,12 +4203,6 @@ declare module 'lodash/zipWith' {
 }
 
 // Filename aliases
-declare module 'lodash/_addMapEntry.js' {
-  declare module.exports: $Exports<'lodash/_addMapEntry'>;
-}
-declare module 'lodash/_addSetEntry.js' {
-  declare module.exports: $Exports<'lodash/_addSetEntry'>;
-}
 declare module 'lodash/_apply.js' {
   declare module.exports: $Exports<'lodash/_apply'>;
 }
@@ -4626,14 +4608,8 @@ declare module 'lodash/_cloneBuffer.js' {
 declare module 'lodash/_cloneDataView.js' {
   declare module.exports: $Exports<'lodash/_cloneDataView'>;
 }
-declare module 'lodash/_cloneMap.js' {
-  declare module.exports: $Exports<'lodash/_cloneMap'>;
-}
 declare module 'lodash/_cloneRegExp.js' {
   declare module.exports: $Exports<'lodash/_cloneRegExp'>;
-}
-declare module 'lodash/_cloneSet.js' {
-  declare module.exports: $Exports<'lodash/_cloneSet'>;
 }
 declare module 'lodash/_cloneSymbol.js' {
   declare module.exports: $Exports<'lodash/_cloneSymbol'>;
@@ -5018,6 +4994,9 @@ declare module 'lodash/_replaceHolders.js' {
 }
 declare module 'lodash/_root.js' {
   declare module.exports: $Exports<'lodash/_root'>;
+}
+declare module 'lodash/_safeGet.js' {
+  declare module.exports: $Exports<'lodash/_safeGet'>;
 }
 declare module 'lodash/_Set.js' {
   declare module.exports: $Exports<'lodash/_Set'>;
