@@ -1,5 +1,5 @@
-// flow-typed signature: 2700fd695eaf09909c98c6cee191ce7c
-// flow-typed version: <<STUB>>/markdown-doctest_v^0.9.1/flow_v0.69.0
+// flow-typed signature: 45e4c7885b7a246673077a9f1a7dd0f8
+// flow-typed version: <<STUB>>/markdown-doctest_v0.9.1/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
