@@ -1,5 +1,5 @@
-// flow-typed signature: 48d93a5cccb4401b447df0031b17287e
-// flow-typed version: <<STUB>>/minimist_v^1.2.0/flow_v0.69.0
+// flow-typed signature: 0eba06168b951ff799dd70974ffc0c1f
+// flow-typed version: <<STUB>>/minimist_v1.2.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
