@@ -1,5 +1,5 @@
-// flow-typed signature: b4d49b43eade2d581899d551f51f5d6f
-// flow-typed version: <<STUB>>/mkdirp_v^0.5.1/flow_v0.69.0
+// flow-typed signature: 2cb8b97960d9f9808a5d068d04a9135e
+// flow-typed version: <<STUB>>/mkdirp_v0.5.1/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
