@@ -1,5 +1,5 @@
-// flow-typed signature: 7b6dd160c34736e67157fb29f7b988e0
-// flow-typed version: <<STUB>>/mocha-in-sauce_v0.0.1/flow_v0.69.0
+// flow-typed signature: a8319e6ecbc29ba411fb926bcd64f66e
+// flow-typed version: <<STUB>>/mocha-in-sauce_v0.0.1/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
