@@ -1,5 +1,5 @@
-// flow-typed signature: fd53c7307894c0f4a6866917d63e2497
-// flow-typed version: <<STUB>>/mocha_v^3.0.2/flow_v0.69.0
+// flow-typed signature: 1eb8ad0d355364a8e1363f22ab0c647b
+// flow-typed version: <<STUB>>/mocha_v5.0.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
