@@ -1,5 +1,5 @@
-// flow-typed signature: 07fbd7cdc9f30d891110ac05a48a6440
-// flow-typed version: <<STUB>>/nanoid_v^1.0.2/flow_v0.69.0
+// flow-typed signature: 2e5d10202182733577dd043d68f351cb
+// flow-typed version: <<STUB>>/nanoid_v^1.0.2/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
