@@ -1,5 +1,5 @@
-// flow-typed signature: 6234280159bb0b8c557267e78aa20f08
-// flow-typed version: <<STUB>>/npm-run-all_v^4.0.2/flow_v0.69.0
+// flow-typed signature: 308914c9ae14c818137df41221196255
+// flow-typed version: <<STUB>>/npm-run-all_v4.1.2/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
