@@ -1,5 +1,5 @@
-// flow-typed signature: 35c47c6135fd3b4461c163886b73d573
-// flow-typed version: <<STUB>>/nyc_v^10.2.0/flow_v0.69.0
+// flow-typed signature: f544de011209d660097a842eaa1f7461
+// flow-typed version: <<STUB>>/nyc_v11.4.1/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
