@@ -1,5 +1,5 @@
-// flow-typed signature: a91a0162c0c538e7acf2a5711be05797
-// flow-typed version: <<STUB>>/opn-cli_v^3.1.0/flow_v0.69.0
+// flow-typed signature: 0d3b424b7fa7876e54288c98b16e950a
+// flow-typed version: <<STUB>>/opn-cli_v3.1.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
