@@ -1,5 +1,5 @@
-// flow-typed signature: e654673b2f03bebba655e0e2cfafb9b8
-// flow-typed version: <<STUB>>/platform_v^1.3.1/flow_v0.69.0
+// flow-typed signature: 8d77eb0e1007f0190223f343284e383c
+// flow-typed version: <<STUB>>/platform_v1.3.5/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
