@@ -1,5 +1,5 @@
-// flow-typed signature: 17bf531edd1e76b75854b91db12ef1b7
-// flow-typed version: <<STUB>>/promise_v^7.1.1/flow_v0.69.0
+// flow-typed signature: 3f30210ddd80609ab349e0e051ba44b0
+// flow-typed version: <<STUB>>/promise_v8.0.1/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
