@@ -1,5 +1,5 @@
-// flow-typed signature: d5fce1fc1f6baa96d4ca2f6275e6f6c5
-// flow-typed version: <<STUB>>/protractor_v^3.1.1/flow_v0.69.0
+// flow-typed signature: d1c89b619114b55f4fc74858ca5d9483
+// flow-typed version: <<STUB>>/protractor_v3.1.1/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
