@@ -1,5 +1,5 @@
-// flow-typed signature: ddd80a4ce56a2a47a60cc34773a3bbfd
-// flow-typed version: <<STUB>>/rollup-plugin-inject_v^2.0.0/flow_v0.69.0
+// flow-typed signature: c4ae7b661bf3815a89adf519a44f70c0
+// flow-typed version: <<STUB>>/rollup-plugin-inject_v2.0.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
