@@ -1,5 +1,5 @@
-// flow-typed signature: c6409fd7d71e03640b08171d6944c860
-// flow-typed version: <<STUB>>/rollup-plugin-node-resolve_v^2.0.0/flow_v0.69.0
+// flow-typed signature: b18a2683524f77a2d4f3a3f02c73edd3
+// flow-typed version: <<STUB>>/rollup-plugin-node-resolve_v2.0.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -15,39 +15,4 @@
 
 declare module 'rollup-plugin-node-resolve' {
   declare module.exports: any;
-}
-
-/**
- * We include stubs for each file inside this npm package in case you need to
- * require those files directly. Feel free to delete any files that aren't
- * needed.
- */
-declare module 'rollup-plugin-node-resolve/dist/rollup-plugin-node-resolve.cjs' {
-  declare module.exports: any;
-}
-
-declare module 'rollup-plugin-node-resolve/dist/rollup-plugin-node-resolve.es' {
-  declare module.exports: any;
-}
-
-declare module 'rollup-plugin-node-resolve/src/empty' {
-  declare module.exports: any;
-}
-
-declare module 'rollup-plugin-node-resolve/src/index' {
-  declare module.exports: any;
-}
-
-// Filename aliases
-declare module 'rollup-plugin-node-resolve/dist/rollup-plugin-node-resolve.cjs.js' {
-  declare module.exports: $Exports<'rollup-plugin-node-resolve/dist/rollup-plugin-node-resolve.cjs'>;
-}
-declare module 'rollup-plugin-node-resolve/dist/rollup-plugin-node-resolve.es.js' {
-  declare module.exports: $Exports<'rollup-plugin-node-resolve/dist/rollup-plugin-node-resolve.es'>;
-}
-declare module 'rollup-plugin-node-resolve/src/empty.js' {
-  declare module.exports: $Exports<'rollup-plugin-node-resolve/src/empty'>;
-}
-declare module 'rollup-plugin-node-resolve/src/index.js' {
-  declare module.exports: $Exports<'rollup-plugin-node-resolve/src/index'>;
 }
