@@ -1,5 +1,5 @@
-// flow-typed signature: 5128e52a941dacd8527949bb14688a50
-// flow-typed version: <<STUB>>/rollup_v0.36.3/flow_v0.69.0
+// flow-typed signature: 7400a2001baddcbfbaa25f0b692f18e9
+// flow-typed version: <<STUB>>/rollup_v0.36.3/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -15,32 +15,4 @@
 
 declare module 'rollup' {
   declare module.exports: any;
-}
-
-/**
- * We include stubs for each file inside this npm package in case you need to
- * require those files directly. Feel free to delete any files that aren't
- * needed.
- */
-declare module 'rollup/dist/rollup.browser' {
-  declare module.exports: any;
-}
-
-declare module 'rollup/dist/rollup.es' {
-  declare module.exports: any;
-}
-
-declare module 'rollup/dist/rollup' {
-  declare module.exports: any;
-}
-
-// Filename aliases
-declare module 'rollup/dist/rollup.browser.js' {
-  declare module.exports: $Exports<'rollup/dist/rollup.browser'>;
-}
-declare module 'rollup/dist/rollup.es.js' {
-  declare module.exports: $Exports<'rollup/dist/rollup.es'>;
-}
-declare module 'rollup/dist/rollup.js' {
-  declare module.exports: $Exports<'rollup/dist/rollup'>;
 }
