@@ -1,5 +1,5 @@
-// flow-typed signature: 1c231857ce0a37e31999cdf4c50cd6a8
-// flow-typed version: <<STUB>>/rx_vlatest/flow_v0.69.0
+// flow-typed signature: ca8b76f858fc424cd93c06dd55c2134d
+// flow-typed version: <<STUB>>/rx_vlatest/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
