@@ -1,5 +1,5 @@
-// flow-typed signature: 4a9410f4b1ca59e0f01e92166df7483a
-// flow-typed version: <<STUB>>/shx_v^0.2.2/flow_v0.69.0
+// flow-typed signature: f4f093e8dfdeb209477e4ebada3101dd
+// flow-typed version: <<STUB>>/shx_v0.2.2/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
