@@ -1,5 +1,5 @@
-// flow-typed signature: aab5858c1eafba493a182114df7355f1
-// flow-typed version: <<STUB>>/sinon-chai_v^2.9.0/flow_v0.69.0
+// flow-typed signature: ce0e5f920cff689d8229e78ee3c19eb3
+// flow-typed version: <<STUB>>/sinon-chai_v2.14.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
