@@ -1,5 +1,5 @@
-// flow-typed signature: 9ca4fca0ae4901b190a53afbfb349ea7
-// flow-typed version: <<STUB>>/source-map-support_v^0.4.0/flow_v0.69.0
+// flow-typed signature: c1b100be021e0b94bcc3605976f60006
+// flow-typed version: <<STUB>>/source-map-support_v0.5.3/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
