@@ -1,5 +1,5 @@
-// flow-typed signature: 827d722b1db8204c8bdad453428f46ce
-// flow-typed version: <<STUB>>/symbol-observable_v1.0.1/flow_v0.69.0
+// flow-typed signature: a84b63b19af7d8b62543fc9139fa9e5c
+// flow-typed version: <<STUB>>/symbol-observable_v1.0.1/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
