@@ -1,10 +1,10 @@
-// flow-typed signature: 29363db6e2e2ce7ea18e86860a8dfb42
-// flow-typed version: <<STUB>>/watch_v^1.0.1/flow_v0.69.0
+// flow-typed signature: efd55a12b98d2a03caa6ada917947133
+// flow-typed version: <<STUB>>/ts-node_v5.0.1/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
  *
- *   'watch'
+ *   'ts-node'
  *
  * Fill this stub out by replacing all the `any` types.
  *
@@ -13,6 +13,6 @@
  * https://github.com/flowtype/flow-typed
  */
 
-declare module 'watch' {
+declare module 'ts-node' {
   declare module.exports: any;
 }
