@@ -1,5 +1,5 @@
-// flow-typed signature: 541fc98a185cc277bc2b6431fce16991
-// flow-typed version: <<STUB>>/tslib_v^1.5.0/flow_v0.69.0
+// flow-typed signature: 94f81470eeac533a1b2e2e6ba94c24ac
+// flow-typed version: <<STUB>>/tslib_v^1.9.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -15,4 +15,25 @@
 
 declare module 'tslib' {
   declare module.exports: any;
+}
+
+/**
+ * We include stubs for each file inside this npm package in case you need to
+ * require those files directly. Feel free to delete any files that aren't
+ * needed.
+ */
+declare module 'tslib/tslib.es6' {
+  declare module.exports: any;
+}
+
+declare module 'tslib/tslib' {
+  declare module.exports: any;
+}
+
+// Filename aliases
+declare module 'tslib/tslib.es6.js' {
+  declare module.exports: $Exports<'tslib/tslib.es6'>;
+}
+declare module 'tslib/tslib.js' {
+  declare module.exports: $Exports<'tslib/tslib'>;
 }
