@@ -1,10 +1,10 @@
-// flow-typed signature: e1171bad33a469e6da7305891160e157
-// flow-typed version: <<STUB>>/sinon_v4.3.0/flow_v0.70.0
+// flow-typed signature: fb47fbae6c7c791167352dbe6a3cddbc
+// flow-typed version: <<STUB>>/tslint-no-unused-expression-chai_v0.0.3/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
  *
- *   'sinon'
+ *   'tslint-no-unused-expression-chai'
  *
  * Fill this stub out by replacing all the `any` types.
  *
@@ -13,6 +13,6 @@
  * https://github.com/flowtype/flow-typed
  */
 
-declare module 'sinon' {
+declare module 'tslint-no-unused-expression-chai' {
   declare module.exports: any;
 }
