@@ -1,5 +1,5 @@
-// flow-typed signature: 94d1c3de1e26f6b49886024c802b079c
-// flow-typed version: <<STUB>>/tslint_v^4.4.2/flow_v0.69.0
+// flow-typed signature: d66157e9f0b99f272b81c0414c1f6b48
+// flow-typed version: <<STUB>>/tslint_v5.9.1/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
