@@ -1,5 +1,5 @@
-// flow-typed signature: 1657ed3d1f414f3318be23c9841ecb0b
-// flow-typed version: <<STUB>>/typescript_v~2.0.6/flow_v0.69.0
+// flow-typed signature: 5514bc5472f14750e555ba0312fd6833
+// flow-typed version: <<STUB>>/typescript_v~2.8.1/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
