@@ -1,5 +1,5 @@
-// flow-typed signature: 9f7dcee6aff1db8e3772a1e094c472a7
-// flow-typed version: <<STUB>>/validate-commit-msg_v^2.14.0/flow_v0.69.0
+// flow-typed signature: 54ba2cd20e34233ee6859c286664d0c9
+// flow-typed version: <<STUB>>/validate-commit-msg_v2.14.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
