@@ -1,5 +1,5 @@
-// flow-typed signature: 289f5c44c323a9fce45d156db4af6dac
-// flow-typed version: <<STUB>>/webpack_v^1.13.1/flow_v0.69.0
+// flow-typed signature: ad5d7b1e0bf91a36a762dd0efac544cc
+// flow-typed version: <<STUB>>/webpack_v1.13.1/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -658,6 +658,10 @@ declare module 'webpack/lib/HotModuleReplacementPlugin' {
   declare module.exports: any;
 }
 
+declare module 'webpack/lib/HotUpdateChunk' {
+  declare module.exports: any;
+}
+
 declare module 'webpack/lib/HotUpdateChunkTemplate' {
   declare module.exports: any;
 }
@@ -963,10 +967,6 @@ declare module 'webpack/lib/performance/SizeLimitsPlugin' {
 }
 
 declare module 'webpack/lib/PrefetchPlugin' {
-  declare module.exports: any;
-}
-
-declare module 'webpack/lib/prepareOptions' {
   declare module.exports: any;
 }
 
@@ -1696,6 +1696,9 @@ declare module 'webpack/lib/HotModuleReplacement.runtime.js' {
 declare module 'webpack/lib/HotModuleReplacementPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/HotModuleReplacementPlugin'>;
 }
+declare module 'webpack/lib/HotUpdateChunk.js' {
+  declare module.exports: $Exports<'webpack/lib/HotUpdateChunk'>;
+}
 declare module 'webpack/lib/HotUpdateChunkTemplate.js' {
   declare module.exports: $Exports<'webpack/lib/HotUpdateChunkTemplate'>;
 }
@@ -1926,9 +1929,6 @@ declare module 'webpack/lib/performance/SizeLimitsPlugin.js' {
 }
 declare module 'webpack/lib/PrefetchPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/PrefetchPlugin'>;
-}
-declare module 'webpack/lib/prepareOptions.js' {
-  declare module.exports: $Exports<'webpack/lib/prepareOptions'>;
 }
 declare module 'webpack/lib/ProgressPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/ProgressPlugin'>;
