@@ -1,5 +1,5 @@
-// flow-typed signature: e0a224b8f9b9cd51beede5116aabeb88
-// flow-typed version: <<STUB>>/xmlhttprequest_v1.8.0/flow_v0.69.0
+// flow-typed signature: 9428ec637e9fda686d5a17f84c684129
+// flow-typed version: <<STUB>>/xmlhttprequest_v1.8.0/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
