@@ -1,5 +1,5 @@
-// flow-typed signature: 6566490155fd64f91aada002d0b44831
-// flow-typed version: <<STUB>>/node-fetch_v^2.1.2/flow_v0.69.0
+// flow-typed signature: 8551dc3c167ad2fe0bb3ce03eccfef34
+// flow-typed version: <<STUB>>/node-fetch_v^2.1.2/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
