@@ -1,5 +1,5 @@
-// flow-typed signature: 9e33d1100c18ac5f0ba04b734bae49a2
-// flow-typed version: <<STUB>>/esm_v^3.0.17/flow_v0.69.0
+// flow-typed signature: 76ac8e944daa18a34f2de7072a41974d
+// flow-typed version: <<STUB>>/esm_v^3.0.26/flow_v0.70.0
 
 /**
  * This is an autogenerated libdef stub for:
